Import PSReadLine change to throw when redirected
A similar change was made to the Source Depot fork of PSReadLine, and
the behavior is depended on in the ConsoleHost. The host always attempts
to load PSReadLine, because a future version of PSReadLine will support
redirected input/output, and until then, PSReadLine must throw so that
the host correctly falls back to Console.ReadLine().